ai_agents/curve_api_agent/requirements.txt
llama-index>=0.9.8
python-dotenv>=1.0.0
openai>=1.3.0
aiohttp>=3.9.1 